+++ examples/printer.rs
@@ -1,6 +1,5 @@
-extern crate openapi;
-
-use std::io::Write;
+use failure::Fail;
+use std::{io::Write, process::exit};
 
 fn main() {
     if let Some(path) = std::env::args().nth(1) {
@@ -17,23 +16,27 @@ fn main() {
                 println!("{}", openapi::to_json(&spec).unwrap());
             }
             Err(e) => {
-
                 let stderr = &mut ::std::io::stderr();
                 let errmsg = "Error writing to stderr";
 
                 writeln!(stderr, "error: {}", e).expect(errmsg);
-
-                for e in e.iter().skip(1) {
-                    writeln!(stderr, "caused by: {}", e).expect(errmsg);
+                for cause in Fail::iter_chain(&e) {
+                    writeln!(
+                        stderr,
+                        "caused by: {} {}",
+                        cause.name().unwrap_or("Error"),
+                        cause
+                    )
+                    .expect(errmsg);
                 }
 
                 // The backtrace is not always generated. Try to run this example
                 // with `RUST_BACKTRACE=1`.
-                if let Some(backtrace) = e.backtrace() {
+                if let Some(backtrace) = e.cause().and_then(|cause| cause.backtrace()) {
                     writeln!(stderr, "backtrace: {:?}", backtrace).expect(errmsg);
                 }
 
-                ::std::process::exit(1);
+                exit(1);
             }
         }
     }