+++ src/lib.rs
@@ -29,46 +29,18 @@
 //! shape and behavior should be consistent and familiar to existing
 //! error_chain users.
 //!
-#[macro_use]
-extern crate error_chain;
-#[macro_use]
-extern crate serde_derive;
-extern crate serde;
-extern crate serde_json;
-extern crate serde_yaml;
-extern crate url;
-extern crate url_serde;
-extern crate semver;
+use serde::{Deserialize, Serialize};
+use std::{fs::File, io::Read, path::Path, result::Result as StdResult};
 
-use std::fs;
-use std::io::Read;
-use std::path::Path;
-
+pub mod error;
 pub mod v2;
 pub mod v3_0;
 
+pub use error::Error;
+
 const MINIMUM_OPENAPI30_VERSION: &str = ">= 3.0";
 
-
-/// errors that openapi functions may return
-pub mod errors {
-    error_chain!{
-        foreign_links {
-            Io(::std::io::Error);
-            Yaml(::serde_yaml::Error);
-            Serialize(::serde_json::Error);
-            SemVerError(::semver::SemVerError);
-        }
-
-        errors {
-            UnsupportedSpecFileVersion(version: ::semver::Version) {
-                description("Unsupported spec file version")
-                display("Unsupported spec file version ({}). Expected {}", version, ::MINIMUM_OPENAPI30_VERSION)
-            }
-        }
-    }
-}
-pub use errors::{Result, ResultExt};
+pub type Result<T> = StdResult<T, Error>;
 
 /// Supported versions of the OpenApi.
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
@@ -91,15 +63,15 @@ pub enum OpenApi {
 }
 
 /// deserialize an open api spec from a path
-pub fn from_path<P>(path: P) -> errors::Result<OpenApi>
+pub fn from_path<P>(path: P) -> Result<OpenApi>
 where
     P: AsRef<Path>,
 {
-    from_reader(fs::File::open(path)?)
+    from_reader(File::open(path)?)
 }
 
 /// deserialize an open api spec from type which implements Read
-pub fn from_reader<R>(read: R) -> errors::Result<OpenApi>
+pub fn from_reader<R>(read: R) -> Result<OpenApi>
 where
     R: Read,
 {
@@ -107,35 +79,30 @@ where
 }
 
 /// serialize to a yaml string
-pub fn to_yaml(spec: &OpenApi) -> errors::Result<String> {
+pub fn to_yaml(spec: &OpenApi) -> Result<String> {
     Ok(serde_yaml::to_string(spec)?)
 }
 
 /// serialize to a json string
-pub fn to_json(spec: &OpenApi) -> errors::Result<String> {
+pub fn to_json(spec: &OpenApi) -> Result<String> {
     Ok(serde_json::to_string_pretty(spec)?)
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::fs::File;
-    use std::io::Write;
-
-    /// Helper function for reading a file to string.
-    fn read_file<P>(path: P) -> String
-    where
-        P: AsRef<Path>,
-    {
-        let mut f = File::open(path).unwrap();
-        let mut content = String::new();
-        f.read_to_string(&mut content).unwrap();
-        content
-    }
+    use pretty_assertions::assert_eq;
+    use std::{
+        fs::{self, read_to_string, File},
+        io::Write,
+    };
 
     /// Helper function to write string to file.
-    fn write_to_file<P>(path: P, filename: &str, data: &str)
-    where
+    fn write_to_file<P>(
+        path: P,
+        filename: &str,
+        data: &str,
+    ) where
         P: AsRef<Path> + std::fmt::Debug,
     {
         println!("    Saving string to {:?}...", path);
@@ -163,12 +130,16 @@ mod tests {
     /// JSON string. The second conversion goes through our `OpenApi`, so the final JSON
     /// string is a representation of _our_ implementation.
     /// By comparing those two JSON conversions, we can validate our implementation.
-    fn compare_spec_through_json(input_file: &Path, save_path_base: &Path) -> (String, String, String) {
+    fn compare_spec_through_json(
+        input_file: &Path,
+        save_path_base: &Path,
+    ) -> (String, String, String) {
         // First conversion:
         //     File -> `String` -> `serde_yaml::Value` -> `serde_json::Value` -> `String`
 
         // Read the original file to string
-        let spec_yaml_str = read_file(&input_file);
+        let spec_yaml_str = read_to_string(&input_file)
+            .unwrap_or_else(|e| panic!("failed to read contents of {:?}: {}", input_file, e));
         // Convert YAML string to JSON string
         let spec_json_str = convert_yaml_str_to_json(&spec_yaml_str);
 
@@ -178,12 +149,13 @@ mod tests {
         // Parse the input file
         let parsed_spec = from_path(&input_file).unwrap();
         // Convert to serde_json::Value
-        let parsed_spec_json: serde_json::Value = serde_json::to_value(parsed_spec).unwrap();
+        let parsed_spec_json = serde_json::to_value(parsed_spec).unwrap();
         // Convert to a JSON string
         let parsed_spec_json_str: String = serde_json::to_string_pretty(&parsed_spec_json).unwrap();
 
         // Save JSON strings to file
-        let api_filename = input_file.file_name()
+        let api_filename = input_file
+            .file_name()
             .unwrap()
             .to_str()
             .unwrap()
@@ -205,8 +177,7 @@ mod tests {
     #[test]
     fn can_deserialize() {
         for entry in fs::read_dir("data/v2").unwrap() {
-            let entry = entry.unwrap();
-            let path = entry.path();
+            let path = entry.unwrap().path();
             // cargo test -- --nocapture to see this message
             println!("Testing if {:?} is deserializable", path);
             from_path(path).unwrap();
@@ -215,43 +186,34 @@ mod tests {
 
     #[test]
     fn can_deserialize_and_reserialize_v2() {
-        let save_path_base: std::path::PathBuf = [
-            "target",
-            "tests",
-            "can_deserialize_and_reserialize_v2",
-        ].iter()
-            .collect();
-        let mut invalid_diffs = Vec::new();
+        let save_path_base: std::path::PathBuf =
+            ["target", "tests", "can_deserialize_and_reserialize_v2"]
+                .iter()
+                .collect();
 
         for entry in fs::read_dir("data/v2").unwrap() {
-            let entry = entry.unwrap();
-            let path = entry.path();
+            let path = entry.unwrap().path();
 
             println!("Testing if {:?} is deserializable", path);
 
             let (api_filename, parsed_spec_json_str, spec_json_str) =
                 compare_spec_through_json(&path, &save_path_base);
 
-            if parsed_spec_json_str != spec_json_str {
-                invalid_diffs.push((api_filename, parsed_spec_json_str, spec_json_str));
-            }
+            assert_eq!(
+                parsed_spec_json_str.lines().collect::<Vec<_>>(),
+                spec_json_str.lines().collect::<Vec<_>>(),
+                "contents did not match for api {}",
+                api_filename
+            );
         }
-
-        for invalid_diff in &invalid_diffs {
-            println!("File {} failed JSON comparison!", invalid_diff.0);
-        }
-        assert!(invalid_diffs.len() == 0);
     }
 
     #[test]
     fn can_deserialize_and_reserialize_v3() {
-        let save_path_base: std::path::PathBuf = [
-            "target",
-            "tests",
-            "can_deserialize_and_reserialize_v3",
-        ].iter()
-            .collect();
-        let mut invalid_diffs = Vec::new();
+        let save_path_base: std::path::PathBuf =
+            ["target", "tests", "can_deserialize_and_reserialize_v3"]
+                .iter()
+                .collect();
 
         for entry in fs::read_dir("data/v3.0").unwrap() {
             let entry = entry.unwrap();
@@ -262,14 +224,12 @@ mod tests {
             let (api_filename, parsed_spec_json_str, spec_json_str) =
                 compare_spec_through_json(&path, &save_path_base);
 
-            if parsed_spec_json_str != spec_json_str {
-                invalid_diffs.push((api_filename, parsed_spec_json_str, spec_json_str));
-            }
+            assert_eq!(
+                parsed_spec_json_str.lines().collect::<Vec<_>>(),
+                spec_json_str.lines().collect::<Vec<_>>(),
+                "contents did not match for api {}",
+                api_filename
+            );
         }
-
-        for invalid_diff in &invalid_diffs {
-            println!("File {} failed JSON comparison!", invalid_diff.0);
-        }
-        assert!(invalid_diffs.len() == 0);
     }
 }